feat: on algos stop the player when path complete
The player will continue following the last movement instruction, so it
will keep using the last movement after it found the diamond.Now I
have added a None instruction, the will make the player stay idle when
it finds the diamond.
This is benifical as it can help generate a path for the next
diamond.