fix: delayed player position value
The x coord value for the player was delayed, so that meant it does not
accuratly represent what gird its in.

This happens because it uses the left hitbox line of player to track the
x grid value. This means if the player is looking left and walking left
it will update the value of its position too quickly.

To fix this I have offset its x position.