refactor: maze is now loaded from a file
The maze map was a 2D array hard coded into main.py, this may not cause
issues but as I make more mazes it will become diffciult to maintain.
So, I have used pickle to generate and store the maze into a object
file, then load up the maze which is needed.