fix: player can not move with a speed less than 1
The players current movement is set to 1 but we need it to move slower
than this if it walks over a 'slow block'. However, pygame rects do not
support decimal movements, it will convert them to ints, so if i did 0.2
it will convert to 0 and the player wont move.

The fix is to use the players center position to update the movement, as
the center argument supports decimal values.