fix: search algos throwing errors
The computer search algorithms were throwing errors after the addition
of the slow blocks.
This is because the walkable maze function did not generate properly
because it did not take into account the additon of the slow blocks,
leaving gaps (zeros) in the matrix. So in most cases it was not possible
to find a path.
I have added a if statment that will check for the slow blocks during
generation.