fix: player stands still for first few seconds
The player will stand still for the first few seconds then start moving
based on the path instructions. At first i thought it was calulating the
path, but based on how small the maze is it should do it almost
instantly.
I looked into this and found that the issue was the palyers coord was
not initally correct, then gets fixed in the draw_animation loop of
AnimationManager. To solve this issue i have correctly gave the inital
coords of the player.