fix: player will keep falling down ladder
The player will keep falling off the ladder, when it almost reaches the
top. This is because i put an conditial in the if statement stating
along the lines off, if the player grid_y adn grid_x values were equal
to 3 then it climb. However, the y_value only reads the top hitbox of
the player so it will keep falling when the top hitbox touches the
bottom line on the top grid of the ladder.

I removed this if conditinol as there was no reason to have it in the
first place.