test: check UCS search chooses the optimal path
In this test there is a short path but it is filled with slow tiles and a
long path but it is filled with normal tiles. In this case we expect the algo
to select the long path because it is the faster of the two.

A new maze was also created for this specfic test case.