test: check every maze generated had a goal
Before there was only a test that will check if maze 1 had a diamond in
it. Now I have updated the test such that it will test every maze has a
diamond in it.

This test is needed as the algos I have coded cannot function properly
without at least one diamond.