feat: improve visuals of higlight grid function
When the highlight grid function was called before, it will highlight
everything at once so the user could not see the order of visited grids. I
have updated this by introducing a sleep function, which slows down the
execution and helps the user see the sequential order of the grids
visited.