test: fix test_bfs_path_find_Script after fail
The test script for test_bfs_can_find_path_in_mid_maze failed because i
slightly changed the logic of the visited list. Before it was actually
adding verticies that were its neighbours and not visited into the path.
I have fixed this issue, but that means the test scripts need to be
updated with new paths to compare with for bfs.