docs: update info on checkstyle README
previous commit, didnt include the intended changes so same comment:
If the user had created there pyhton enviorment in the product
directory, then flake 8 will check the contents of this file leading to
alot of errors. To fix this i added an exclude option in the README to
avoid this.