fix: random algo kept crashing
The random algo was crashing because firstly it was in an inifinite loop
trying to generate a path, but the random algo dosent have a path to
generate. I put an condition to handle this.

Secondly, it didnt have this function update_diamond_list(). To fix this
I move the function to the root Computer class, so its accessible to all
computer types and not just greedy.