fix: agent would keep climbing even when blocked
Sometimes the agent will try and climb a ladder when there was a block
on top of it.
Thius happens at the bottom of the ladder, rather than the top.
This is because the agent is not properly aligned in the ladder grid, in
other words it starts climbing too early. Half the agent hitbox will be
under a block as a result, preventing it from climbing.

To fix this, before the agent starts climbing. I will force the agent
coords to be aligned in the ladder grid.