fix: agent stops moving after score of 5
On any non competitive algos, the agent will stop moving after a score
of 5. This is bacause I had some POC code for minimax which I never
ended up using but affected the non comp algos. It basically stops the
algo, if it cant traverse a path in a given time.